+++ backend/safe-service/src/main/java/com/safeservice/domain/report/repository/ReportRepositoryImpl.java
@@ -33,7 +33,6 @@ public class ReportRepositoryImpl implements ReportRepositoryCustom{
     public List<Report> searchAll() {
         return jpaQueryFactory.selectFrom(report)
                 .leftJoin(report.reportFiles, reportFile).fetchJoin()
-                .where(reportFile.reportFileType.eq(ReportFileType.RESIZE))
                 .fetch();
     }
